# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/book": {
            "get": {
                "description": "根据ID获取书籍信息",
                "tags": [
                    "book"
                ],
                "summary": "获取书籍",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "书籍ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.BookGetRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/book/create": {
            "post": {
                "description": "创建新的书籍",
                "tags": [
                    "book"
                ],
                "summary": "创建书籍",
                "parameters": [
                    {
                        "description": "书籍信息",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BookCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.BookCreateRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.BookCreateReq": {
            "type": "object",
            "required": [
                "author",
                "name",
                "price"
            ],
            "properties": {
                "author": {
                    "description": "作者",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1,
                    "example": "张三"
                },
                "name": {
                    "description": "书籍名称",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1,
                    "example": "Go语言编程"
                },
                "price": {
                    "description": "价格",
                    "type": "number",
                    "minimum": 0.01,
                    "example": 100
                }
            }
        },
        "dto.BookCreateRes": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "书籍id",
                    "type": "integer"
                }
            }
        },
        "dto.BookGetRes": {
            "type": "object",
            "properties": {
                "author": {
                    "description": "作者",
                    "type": "string"
                },
                "create_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "书籍ID",
                    "type": "integer"
                },
                "name": {
                    "description": "书籍名称",
                    "type": "string"
                },
                "price": {
                    "description": "价格",
                    "type": "number"
                }
            }
        },
        "enum.Code": {
            "type": "integer",
            "enum": [
                200,
                400,
                401,
                403,
                404,
                500
            ],
            "x-enum-varnames": [
                "OK",
                "BadRequest",
                "Unauthorized",
                "Forbidden",
                "HttpNotFound",
                "InternalError"
            ]
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/enum.Code"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}